+++ docker-publish.sh
@@ -1,8 +1,5 @@
 #!/bin/bash
-set -eu
-
-DOCKER_USERNAME="$1"
-DOCKER_PASSWORD="$2"
+set -eux
 
 echo "$DOCKER_PASSWORD" | docker login -u "$DOCKER_USERNAME" --password-stdin
 docker tag ${DOCKER_IMAGE_NAME}:${DOCKER_IMAGE_TAG} ${DOCKER_IMAGE_NAME}:${TRAVIS_TAG}-${TRAVIS_BUILD_NUMBER}